Adds confirmation before deleting a device

+++ src/app/pages/device-control/device-control.component.ts
@@ -58,28 +58,41 @@ export class DeviceControlComponent implements OnInit {
       })
     );
   }
-  delete() {
-    this.subscription.add(
-      this.api.deleteMyDevice(this.myDevice.deviceMeta.uuid).subscribe(
-        async (res) => {
-          await swal({
-            title: `${this.myDevice.name} has been deleted from your device list.`,
-            type: "success",
-            timer: 2000,
-            showConfirmButton: false,
-            showCloseButton: false,
-          });
-          this.onBackButtonClick();
-        },
-        async (err) => {
-          await swal({
-            title: `Error!!!`,
-            text: `Could not delete ${this.myDevice.name} from your device list`,
-            type: "error",
-          });
-        }
-      )
-    );
+  async delete() {
+    let confirm = await swal({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      return result.value;
+    });
+    if (confirm) {
+      this.subscription.add(
+        this.api.deleteMyDevice(this.myDevice.deviceMeta.uuid).subscribe(
+          async (res) => {
+            await swal({
+              title: `${this.myDevice.name} has been deleted from your device list.`,
+              type: "success",
+              timer: 2000,
+              showConfirmButton: false,
+              showCloseButton: false,
+            });
+            this.onBackButtonClick();
+          },
+          async (err) => {
+            await swal({
+              title: `Error!!!`,
+              text: `Could not delete ${this.myDevice.name} from your device list`,
+              type: "error",
+            });
+          }
+        )
+      );
+    }
   }
   getHeaderButtons(): HeaderButton[] {
     return this.loading ? [] : this.headerButtons;
